G78 button to LR0

+++ runahead.c
@@ -1558,7 +1558,7 @@ static INLINE void preempt_input_poll(preempt_t *preempt,
       runloop_state_t *runloop_st, settings_t *settings)
 {
    size_t p;
-   int16_t joypad_state;
+   int32_t joypad_state;
    retro_input_state_t state_cb   = input_driver_state_wrapper;
    unsigned max_users             = settings->uints.input_max_users;
